Fixing RandomChar
Fixing the characters component from the server, and correcting the code.
I just added destructurization, due to the fact that I did not notice the destructurization, the data was not shown. I also used componentDidMount() to fix a bug that occurred due to the fact that I call the updateChar method in the constructor. Fixing the characters component from the server, and correcting the code.
I just added destructurization, due to the fact that I did not notice the destructurization, the data was not shown. I also used componentDidMount() to fix a bug that occurred due to the fact that I call the update Char method in the constructor. I also made the characters change every 3 seconds